% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/converge_estimate.R
\name{converge.estimate}
\alias{converge.estimate}
\title{Get convergence estimate}
\usage{
converge.estimate(text, step.size, max.length = length(text),
  every.word = 10, downsampling.rate = 5, verbose = TRUE)
}
\arguments{
\item{text}{Character vector. Preferably this should be read with
\code{readdile} to take care of preprocessing. Note
that internally this uses a regular expression for the matching
so in non-ASCII texts it might be unstable. In those cases it
is recommended to \code{normalize} the text beforehand.}

\item{step.size}{A numeric scalar. Every how many words should an
estimate be computed? Warning: this can increase the
computation time.}

\item{max.length}{A numeric scalar. How many elements from
\code{text} should be read? Defaults to the length of the text.}

\item{every.word}{A numeric scalar.}

\item{downsampling.rate}{A numeric scalar.}

\item{verbose}{Boolean. Whether or not to display percentage done.}
}
\value{
`converge.estimate` returns an object of `class` `"Converge"`.

An object of the class `"Converge"` has the following attributes:

convergence: A data frame. 
min.criterion: The minimum SD value found.
convergence.criterion: A data frame
}
\description{
\code{converge.estimate} is used to find the minimum size of text
needed to get a stable entropy rate.
}
\examples{
data(deuparl)
ce <- converge.estimate(deuparl, step.size = 1000, max.length = 10000, every.word = 10, downsampling.rate = 5, verbose = TRUE)
}
