% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\docType{data}
\name{deuparl}
\alias{deuparl}
\title{First 50000 words of the German part if the Europarl Corpus}
\format{A character vector with 50000 elements}
\source{
\url{http://www.statmt.org/europarl/}
}
\usage{
deuparl
}
\description{
A dataset containing the first 50000 words of the German part of
the European Parliament Proceedings Parallel Corpus.
}
\keyword{datasets}
