% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/converge_estimate.R
\name{get.estimate}
\alias{get.estimate}
\title{Get single convergence estimate}
\usage{
get.estimate(text, max.length = length(text), every.word = 10)
}
\arguments{
\item{text}{Character vector. Preferably this should be read with
\code{readdile} to take care of preprocessing. Note
that internally this uses a regular expression for the matching
so in non-ASCII texts it might be unstable. In those cases it
is recommended to \code{normalize} the text beforehand.}

\item{max.length}{A numeric scalar. How many elements from
\code{text} should be read? Defaults to the length of the text.}
}
\value{
A numeric scalar. This is the entropy rate on the given
    chunk of text.
}
\description{
\code{get.estimate} is used to return a single entropy rate
estimate for a given text.
}
\examples{
data(deuparl)
est <- get.estimate(deuparl, max.length = 10000)
}
