% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{PreprocessText}
\alias{PreprocessText}
\title{Preprocess character vectors}
\usage{
PreprocessText(text, lower = FALSE, bow = TRUE)
}
\arguments{
\item{text}{A character vector. This contains the text as returned
by \code{scan}.}

\item{lower}{Boolean. Whether or not to lowercase all words.}

\item{bow}{Boolean. Whether or not to substitute each word with an
ID tag (useful for non-ASCII texts)}
}
\value{
A character vector.
}
\description{
This provides some elementary preprocessing for a read character
vector such as lowercasing and bag-of-words normalization. The bow
normalization step substitutes each element of the vector with a
numeric value (its ID). This can be quite useful in non-ASCII texts
or texts containing words with boundary symbols where the regular
expression can fail.
}
\examples{
txt <- c("This", "is", "a", "Sentence", "containing", "UPPERCASE", "lowercase", "and", "sy.mb'ols")
txt.norm <- PreprocessText(txt, lower = TRUE, bow = TRUE)
}
\seealso{
\code{\link{tolower}}
}
