% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{read.file}
\alias{read.file}
\title{Read character vector from file}
\usage{
read.file(text.file, max.length = -1, lower = FALSE, bow = FALSE, ...)
}
\arguments{
\item{text.file}{A character vector. The path to the text file}

\item{max.length}{A numeric. The maximum length of space delimited
units to be read (defaults to the entire file).}

\item{lower}{Boolean. Whether or not to lowercase all words.}

\item{bow}{Boolean. Whether or not to substitute each word with an
ID tag (useful for non-ASCII texts)}

\item{...}{arguments to be passed to \code{scan}.}
}
\description{
\code{read.file} reads a file word by word and returns a unicode
character vector.
}
\details{
The \code{read.file} helper function reads a character vector from
a file up to \code{max.length} calling \code{\link{PreprocessText}}
if nececssary to carry out any preprocessing steps.
}
\examples{
text <- read.file("mycorpus.txt", max.length = 1000, lower = True, bow = TRUE)
}
\seealso{
\code{\link{scan}}, \code{\link{PreprocessText}}
}
