% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClass.R, R/AllMethods.R
\docType{class}
\name{StabilizeEntropy-class}
\alias{StabilizeEntropy-class}
\alias{StabilizeEntropy}
\alias{StabilizeEntropy-class}
\alias{StabilizeEntropy-class}
\alias{StabilizeEntropy-class}
\alias{StabilizeEntropy-class}
\alias{StabilizeEntropy}
\alias{Stabilize,StabilizeEntropy-method}
\alias{StabilizeEntropy,StabilizeEntropy-method}
\alias{StabilizeCriterion,StabilizeEntropy-method}
\alias{get.stabilization,StabilizeEntropy-method}
\alias{get.criterion,StabilizeEntropy-method}
\title{StabilizeEntropy class.}
\usage{
StabilizeCriterion(object, method, rate)

Stabilize(object, verbose, random)

get.stabilization(object)

get.criterion(object)

\S4method{initialize}{StabilizeEntropy}(.Object, ..., text, step.size,
  max.length, every.word)

\S4method{Stabilize}{StabilizeEntropy}(object, verbose)

\S4method{StabilizeCriterion}{StabilizeEntropy}(object, method, rate)

\S4method{get.stabilization}{StabilizeEntropy}(object)

\S4method{get.criterion}{StabilizeEntropy}(object)
}
\description{
This is mainly a delegating class encapsulating the logic of the
simulations. Concretely, it estimates the _entropy rate_ for
different text sizes by repeatedly calling GetSingleEstimate then
uses those estimate to find the stabilization point. It also holds
helper printing and plotting methods.

Method StabilizeCriterion.

Method Stabilize.

Method get.stabilization.

Method get.criterion.

Constructor method of StabilizeEntropy Class.
}
\section{Slots}{

\describe{
\item{\code{text}}{a character vector. A character vector to use as corpus}

\item{\code{step.size}}{a numeric scalar. The size of each chunk of corpus
to consider}

\item{\code{max.length}}{a numeric scalar. The total size of the text to
consider}

\item{\code{every.word}}{a numeric scalar. Every how many words should it
compute the entropy rate (reduces computation time)}

\item{\code{cache.obj}}{an environment. This is used to memoize already
encountered chunks in the \code{\link{stabilize.estimate}}}

\item{\code{min.criterion}}{a numeric scalar. The criterion to use for
stabilization.}

\item{\code{stabilization}}{a data.frame.}

\item{\code{stabilization.criterion}}{a data.frame.}
}}

\author{
Dimitris Alikaniotis
}
