% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClass.R
\docType{class}
\name{stabilizeentropy}
\alias{stabilizeentropy}
\alias{StabilizeEntropy}
\title{StabilizeEntropy class}
\description{
This is mainly a delegating class encapsulating the logic of the
simulations. Concretely, it estimates the _entropy rate_ for
different text sizes by repeatedly calling GetSingleEstimate then
uses those estimate to find the stabilization point. It also holds
helper printing and plotting methods.
}
\section{Slots}{

\describe{
\item{\code{text}}{(character): A character vector to use as corpus}

\item{\code{step.size}}{(numeric): The size of each chunk of corpus to
consider}

\item{\code{max.length}}{(numeric): The total size of the text to consider}

\item{\code{every.word}}{(numeric): Every how many words should it compute
the entropy rate (reduces computation time)}
}}

\author{
Dimitris Alikaniotis
}
