% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stabilize_estimate.R
\name{get.estimate}
\alias{get.estimate}
\title{Get single stabilization estimate}
\usage{
get.estimate(text, max.length = length(text), every.word = 10)
}
\arguments{
\item{text}{A character vector. Preferably this should be read with
\code{\link{read.file}} to take care of preprocessing. Note
that internally this uses a regular expression for the matching
so in non-ASCII texts it might be unstable. In those cases it
is recommended to \code{\link{normalize}} the text beforehand.}

\item{max.length}{A numeric scalar. How many elements from
\code{text} should be read? Defaults to the length of the text.}

\item{every.word}{A numeric scalar. "every.word=1" specifies that
each word token should be used for estimation. To speed up
processing only every 2nd, 3rd, xth word token could be used.
Hence, every.word can be assigned any integer between 1 and the
step size.}
}
\value{
A numeric scalar. This is the entropy rate on the given
    chunk of text.
}
\description{
\code{get.estimate} is used to return a single entropy rate
estimate for a given text.
}
\examples{
data(deuparl)
est <- get.estimate(deuparl, max.length = 10000, every.word = 10)
}
\seealso{
\code{\link{PreprocessText}}, \code{\link{read.file}}, \code{\link{IsSubstring}}
}
