% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{IsSubstring}
\alias{IsSubstring}
\title{Tests whether `subvec` is a proper substring of `fullvec`.}
\usage{
IsSubstring(subvec, fullstr)
}
\description{
Tests whether `subvec` is a proper substring of `fullvec`.
}
